MPIEvaluator: Fix logging by explicitly passing logging level
Pass the log level explicitly to the run_experiment_mpi() function.

The level passed to log_to_stderr isn't available when the _logger object is initiated earlier. To migrate this problem, we explicitly pass it.

The initializer first took care of that, but since that wouldn't allow multiple consecutive calls to the MPIEvaluator it had to be removed again.

Now there are two sets of options:
- ema_logging.log_to_stderr(level=n, set_root_logger_levels=True) --> Everything level n
- ema_logging.log_to_stderr(level=n, set_root_logger_levels=False) --> Evaluator module level n, internal MPI / EMA process level 10 (default)

When level=10, set_root_logger_levels True or False doesn't matter.

Might have performance impact by setting up a logging.basicConfig for every experiment.